Split script to script_vm and script
Move gNullScriptPtr to the script_vm file.
