+++ src/scrcmd.c
@@ -70,10 +70,6 @@ extern const u8 *gStdScripts_End[];
 
 static void CloseBrailleWindow(void);
 
-// This is defined in here so the optimizer can't see its value when compiling
-// script.c.
-void *const gNullScriptPtr = NULL;
-
 static const u8 sScriptConditionTable[COMPARISON_OPERATORS_COUNT][3] =
 {
 //                              <  =  >
